+++ src/propertyFields/collectionData/collectionDataItem/ICollectionDataItemProps.ts
@@ -4,10 +4,13 @@ export interface ICollectionDataItemProps {
   fields: ICustomCollectionField[];
   index?: number;
   item?: any;
+  sortingEnabled?: boolean;
+  totalItems?: number;
 
   fAddItem?: (item: any) => void;
   fAddInCreation?: (item: any) => void;
   fUpdateItem?: (idx: number, item: any) => void;
   fDeleteItem?: (idx: number) => void;
   fValidation?: (idx: number, isValid: boolean) => void;
+  fOnSorting?: (oldIdx: number, newIdx: number) => void;
 }